kinect1: change processedFrameBuffer color type to FLOAT32

+++ orx-jvm/orx-kinect-v1/src/main/kotlin/Kinect1.kt
@@ -213,7 +213,7 @@ class Kinect1 : Kinect, Extension {
                 resolution.x,
                 resolution.y,
                 format = ColorFormat.R,
-                type = ColorType.FLOAT16 // in the future we might want to choose the precision here
+                type = ColorType.FLOAT32
             ).also {
                 it.filter(MinifyingFilter.LINEAR, MagnifyingFilter.LINEAR)
             }
